Did stuff

+++ TowerDefense/Model/Enemy/BaseEnemy.cs
@@ -8,10 +8,11 @@ namespace TowerDefense.Model.Enemies
 {
     public abstract class Enemy
     {
-       public int Health;
-       public int Speed;
-       public int x;
-       public int y;
-       public int Goldgiven;
+       
+       public short Health;
+       public short Speed;
+       public short x;
+       public short y;
+       public short Goldgiven;
     }
 }
